removing unnecessary utils import as relativeTime is only used in one place by one file.
